refactor(code smells/duplication): reduce code smell and duplication from previous branch
- create CommentService and BookmarkService files to abstract methods from ArticleService file
- create bookmarks and comments resolver to abstract methods from articles resolver file
