+++ server/resolvers/index.js
@@ -1,9 +1,13 @@
 import userResolver from './users';
 import articleResolver from './articles';
+import bookmarkResolver from './bookmarks';
+import commentResolver from './comments';
 
 const rootResolver = {
   ...userResolver,
-  ...articleResolver
+  ...articleResolver,
+  ...bookmarkResolver,
+  ...commentResolver
 };
 
 export default rootResolver;
